# Speedup screenshots and update dependencies