--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -1,14 +1,11 @@
 module.exports = {
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es2021": true,
-        "node": true
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-        "ecmaVersion": 12
-    },
-    "rules": {
-    }
+  'env': {
+    'browser': true,
+    'commonjs': true,
+    'es2021': true,
+    'node': true
+  },
+  'extends': 'eslint:recommended',
+  'parserOptions': { 'ecmaVersion': 12 },
+  'rules': {'indent': ['error', 2]}
 };