--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -14,6 +14,9 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v2
 
+      - name: Setup firefox
+        uses: browser-actions/setup-firefox@latest
+
       - name: Install modules
         run: npm ci --no-audit