**IMPORTANT**: Github are updating ubuntu-latest to point to ubuntu-22.04, which does not, as of yet, have firefox installed. That's why the explicit setup of firefox. (If you are wondering why so many builds)

* bumped the version in `package.json` to match the released one.
* Removed sleep time after `driver.get(url)` because the default behaviour is to wait for `window.load`
  * Note: if [default `pageLoadStrategy` changes](https://www.selenium.dev/documentation/webdriver/drivers/options/#pageloadstrategy), pastshots should be updated
* sprinkled a lot of timers here and there to measure various things
  * Note: commented out because the timers might be useful. I didn't want to add `--debug` flag
* Updated dependencies for almost everything, because I couldn't install on M macs 
  * `setSize` no longer exists and is `setRect` instead
  * looksSame has a different signature
  * `fs.mkdirSync` supports recursive

Here are some metrics from said timers:

```
driver.setSize: 1.647s

Loading http://localhost:8081/example/pages/_first.html...
  driver.get url: 417.657ms
  driver.findElement: 27.404ms (viewport)
  driver.takeScreenshot: 31.509ms
  looksSame.compare: 35.562ms
  runTest: 512.528ms

Loading http://localhost:8081/example/pages/buttons.html...
  driver.get url: 45.072ms
  driver.findElement: 16.75ms (viewport)
  driver.takeScreenshot: 35.363ms
  looksSame.compare: 43.858ms
  runTest: 141.742ms

Loading http://localhost:8081/example/pages/more-pages/icons.html...
  driver.get url: 30.964ms
  driver.findElement: 10.687ms (viewport)
  driver.takeScreenshot: 20.916ms
  looksSame.compare: 32.258ms
  runTest: 95.383ms

driver quit: 1.031s

all tests: 1.902s

pastshots.capture: 3.550s

pastshots: 3.718s
```

There are some one time, but expensive operations like `driver.setSize` and `driver.quit`. The first file to be loaded is the slowest, because there is no cache. Optimizing the images is relatively slow (about 250ms), but since it's asynchronous, we don't wait for it to complete, before we go to the next test. It only adds wait for the last test.